Update data handling of Callhome
The new way of handling Callhome is to download the datasets (Talkbank and the English extension), set the right paths in the pipe.sh script, then run it to preprocess the dataset. Then you can use the CallhomePipeline to load a split as you need, which will load the entire text dataset into memory (it's small enough). You can then use it for downstream processing - it provides spanish/english pairs of text, alongside audio and translations if needed.